+++ app/client/src/components/shared/MapWidgets.tsx
@@ -325,14 +325,6 @@ const resizeHandleStyles = css`
   position: absolute;
   right: 0;
   bottom: 0;
-
-  .fa-rotate-45 {
-    transform: rotate(45deg);
-  }
-
-  .fa-rotate-315 {
-    transform: rotate(315deg);
-  }
 `;
 
 /*
@@ -1952,6 +1944,9 @@ function ShowSelectedUpstreamWatershed({
     if (!view) return;
     const mapClickHandler = view.on('click', async (ev) => {
       if (!selectionActive) return;
+
+      ev.stopPropagation();
+
       let res;
       try {
         res = await handleHucSelection(ev);