+++ app/client/src/components/shared/WaterbodyList.tsx
@@ -3,6 +3,7 @@
 import { useContext } from 'react';
 import { css } from '@emotion/react';
 // components
+import { WarningIcon } from 'components/shared/Icons';
 import LoadingSpinner from 'components/shared/LoadingSpinner';
 import WaterbodyIcon from 'components/shared/WaterbodyIcon';
 import WaterbodyInfo from 'components/shared/WaterbodyInfo';
@@ -179,7 +180,7 @@ function WaterbodyList({ waterbodies, title, fieldName }: Props) {
                     <>
                       <br />
                       <span css={noMapDataWarningStyles}>
-                        <i className="fas fa-exclamation-triangle" />
+                        <WarningIcon />
                         <strong>[Waterbody not visible on map.]</strong>
                       </span>
                     </>