+++ app/client/src/utils/streamgages.ts
@@ -1,5 +1,6 @@
-import Graphic from '@arcgis/core/Graphic';
+import * as containsOperator from '@arcgis/core/geometry/operators/containsOperator.js';
 import FeatureLayer from '@arcgis/core/layers/FeatureLayer';
+import Graphic from '@arcgis/core/Graphic';
 import Point from '@arcgis/core/geometry/Point';
 import SimpleMarkerSymbol from '@arcgis/core/symbols/SimpleMarkerSymbol';
 import SimpleRenderer from '@arcgis/core/renderers/SimpleRenderer';
@@ -126,7 +127,7 @@ function useUpdateData() {
       usgsSiteTypes,
       usgsStaParameters,
       controller.signal,
-      huc12,
+      (hucBoundaries?.geometry as __esri.Polygon) ?? null,
     ).then((data) => {
       setHucData(data);
     });
@@ -275,7 +276,7 @@ async function fetchAndTransformData(
   usgsSiteTypes: ConfigFiles['usgsSiteTypes'],
   usgsStaParameters: UsgsStaParameter[],
   abortSignal: AbortSignal,
-  huc12: string | null = null,
+  huc12Boundaries: __esri.Polygon | null = null,
   additionalData?: UsgsStreamgageAttributes[] | null,
 ) {
   dispatch({ type: 'pending', id: fetchedDataId });
@@ -299,7 +300,12 @@ async function fetchAndTransformData(
   });
 
   const responses = await Promise.all([
-    fetchMonitoringLocations(monLocIdSet, services, abortSignal, huc12),
+    fetchMonitoringLocations(
+      monLocIdSet,
+      services,
+      abortSignal,
+      huc12Boundaries,
+    ),
     fetchDaily(monLocIdSet, services, abortSignal),
     fetchPrecipitation(monLocIdSet, services, abortSignal),
   ]);
@@ -561,7 +567,7 @@ function fetchMonitoringLocations(
   monitoringLocations: Set<string>,
   servicesData: ServicesData,
   abortSignal: AbortSignal,
-  huc12: string | null = null,
+  huc12Boundaries: __esri.Polygon | null = null,
 ): Promise<FetchState<UsgsMonitoringLocationData>> {
   let url =
     servicesData.usgs.monitoringLocations +
@@ -569,9 +575,7 @@ function fetchMonitoringLocations(
     `&limit=10000` +
     `&properties=agency_code,monitoring_location_number,monitoring_location_name,site_type,site_type_code`;
 
-  if (huc12) url += `&hydrologic_unit_code=${huc12}`;
-
-  return fetchPost(
+  return fetchPost<UsgsMonitoringLocationData>(
     url,
     {
       op: 'in',
@@ -589,7 +593,22 @@ function fetchMonitoringLocations(
     .then((res) => {
       return {
         status: 'success',
-        data: res,
+        data: {
+          ...res,
+          features: res.features.filter((feat) => {
+            const geometry = new Point({
+              longitude: feat.geometry.coordinates[0],
+              latitude: feat.geometry.coordinates[1],
+              spatialReference: {
+                wkid: 102100,
+              },
+            });
+            return (
+              !huc12Boundaries ||
+              containsOperator.execute(huc12Boundaries, geometry)
+            );
+          }),
+        },
       } as FetchSuccessState<UsgsMonitoringLocationData>;
     })
     .catch(handleFetchError);